+++ libs/common/c-data/src/lib/core/enums/sort-direction.enum.ts
@@ -1,0 +1,4 @@
+export enum sortDirection {
+  ascend = 'ASC',
+  descend = 'DESC',
+}